+++ lib/mode/dialog_report_mode.dart
@@ -8,34 +8,35 @@ class DialogReportMode extends ReportMode {
     _showDialog(report, context);
   }
 
-  _showDialog(Report report, BuildContext context) async {
+  Future _showDialog(Report report, BuildContext context) async {
     await Future.delayed(Duration.zero);
     showDialog(
-        context: context,
-        builder: (BuildContext build) {
-          return AlertDialog(
-            title: Text(localizationOptions.dialogReportModeTitle),
-            content: Text(localizationOptions.dialogReportModeDescription),
-            actions: <Widget>[
-              FlatButton(
-                child: Text(localizationOptions.dialogReportModeAccept),
-                onPressed: () => _acceptReport(context, report),
-              ),
-              FlatButton(
-                child: Text(localizationOptions.dialogReportModeCancel),
-                onPressed: () => _cancelReport(context, report),
-              ),
-            ],
-          );
-        });
+      context: context,
+      builder: (BuildContext build) {
+        return AlertDialog(
+          title: Text(localizationOptions.dialogReportModeTitle),
+          content: Text(localizationOptions.dialogReportModeDescription),
+          actions: <Widget>[
+            FlatButton(
+              child: Text(localizationOptions.dialogReportModeAccept),
+              onPressed: () => _onAcceptReportClicked(context, report),
+            ),
+            FlatButton(
+              child: Text(localizationOptions.dialogReportModeCancel),
+              onPressed: () => _onCancelReportClicked(context, report),
+            ),
+          ],
+        );
+      },
+    );
   }
 
-  _acceptReport(BuildContext context, Report report) {
+  void _onAcceptReportClicked(BuildContext context, Report report) {
     super.onActionConfirmed(report);
     Navigator.pop(context);
   }
 
-  _cancelReport(BuildContext context, Report report) {
+  void _onCancelReportClicked(BuildContext context, Report report) {
     super.onActionRejected(report);
     Navigator.pop(context);
   }