+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.3.4
+* Fixed typos by Rob Halff (https://github.com/rhalff)
+* [BREAKING CHANGE] Changed customTitle to title and customDescription to description in error widget.
+* Added support for using error widget in small widgets.
+* Refactored classes and added proper input handling.
+* Fixed HttpHandler headers.
+* Fixed AndroidX support.
+
+## 0.3.3
+* Updated dependencies
+* Prepare for 1.0.0 version of sensors and package_info. ([dart_lsc](http://github.com/amirh/dart_lsc))
+
 ## 0.3.2
 * Removed invalid import from error widget